Rework makefile option to make more robust
