+++ setupext/__init__.py
@@ -2,7 +2,6 @@ from . import utils
 from . import dist
 from . import platform
 from . import build_ext
-from . import build_makefile
 from . import build_java
 from . import build_thunk
 from . import test_java
